Imported WebKit commit 499c84c99aa98e9870fa7eaa57db476c6d160d46 (http://svn.webkit.org/repository/webkit/trunk@119200)
Weekly update :). Particularly relevant changes for Qt are the use of the WebCore image decoders and direct usage
of libpng/libjpeg if available in the system.

+++ Source/WebCore/inspector/PageRuntimeAgent.cpp
@@ -151,7 +151,7 @@ void PageRuntimeAgent::unmuteConsole()
 void PageRuntimeAgent::notifyContextCreated(const String& frameId, ScriptState* scriptState, SecurityOrigin* securityOrigin, bool isPageContext)
 {
     ASSERT(securityOrigin || isPageContext);
-    long executionContextId = injectedScriptManager()->injectedScriptIdFor(scriptState);
+    int executionContextId = injectedScriptManager()->injectedScriptIdFor(scriptState);
     String name = securityOrigin ? securityOrigin->toString() : "";
     m_frontend->isolatedContextCreated(ExecutionContextDescription::create()
         .setId(executionContextId)
